Create linearSearch
Write a C++ function for linear search that searches for a given value in an array. The function should take an integer array, the size of the array, and the value to be searched as parameters. If the value is found in the array, the function should return the index where it is located; otherwise, it should return -1. Additionally, provide a simple example in C++ demonstrating the usage of this function with a sample array and a value to search.